refactor: created new package resources for resources handling, added new abstract class for resources to outsource common code, specific services for resources were created
