+++ src/main/java/de/fraunhofer/isst/configmanager/api/controller/ResourceContractController.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import de.fraunhofer.iais.eis.ContractOffer;
 import de.fraunhofer.iais.eis.ids.jsonld.Serializer;
 import de.fraunhofer.isst.configmanager.api.ResourceContractApi;
-import de.fraunhofer.isst.configmanager.api.service.ResourceService;
+import de.fraunhofer.isst.configmanager.api.service.resources.ResourceContractService;
 import de.fraunhofer.isst.configmanager.connector.clients.DefaultResourceClient;
 import de.fraunhofer.isst.configmanager.model.usagecontrol.Pattern;
 import de.fraunhofer.isst.configmanager.util.ValidateApiInput;
@@ -33,15 +33,15 @@ import java.net.URI;
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
 public class ResourceContractController implements ResourceContractApi {
 
-    transient ResourceService resourceService;
+    transient ResourceContractService resourceContractService;
     transient Serializer serializer;
     transient DefaultResourceClient client;
 
     @Autowired
-    public ResourceContractController(final ResourceService resourceService,
+    public ResourceContractController(final ResourceContractService resourceContractService,
                                       final Serializer serializer,
                                       final DefaultResourceClient client) {
-        this.resourceService = resourceService;
+        this.resourceContractService = resourceContractService;
         this.serializer = serializer;
         this.client = client;
     }
@@ -57,7 +57,7 @@ public class ResourceContractController implements ResourceContractApi {
         log.info(">> GET /resource/contract resourceId: " + resourceId);
         ResponseEntity<String> response;
 
-        final var contractOffer = resourceService.getResourceContract(resourceId);
+        final var contractOffer = resourceContractService.getResourceContract(resourceId);
         if (contractOffer != null) {
             try {
                 response = ResponseEntity.ok(serializer.serialize(contractOffer));
@@ -101,7 +101,7 @@ public class ResourceContractController implements ResourceContractApi {
 
                         final var clientResponse = client.updateResourceContract(resourceId.toString(), contractJson);
 
-                        resourceService.updateResourceContractInAppRoute(resourceId, contractOffer);
+                        resourceContractService.updateResourceContractInAppRoute(resourceId, contractOffer);
 
                         jsonObject.put("connectorResponse", clientResponse);
                         response = ResponseEntity.ok(jsonObject.toJSONString());
@@ -142,7 +142,7 @@ public class ResourceContractController implements ResourceContractApi {
         } else {
             ContractOffer contractOffer = null;
             try {
-                contractOffer = resourceService.getContractOffer(pattern, contractJson);
+                contractOffer = resourceContractService.getContractOffer(pattern, contractJson);
             } catch (JsonProcessingException e) {
                 log.error(e.getMessage());
             }
@@ -158,7 +158,7 @@ public class ResourceContractController implements ResourceContractApi {
 
                     final var connectorResponse = client.updateResourceContract(resourceId.toString(), contract);
 
-                    resourceService.updateResourceContractInAppRoute(resourceId, contractOffer);
+                    resourceContractService.updateResourceContractInAppRoute(resourceId, contractOffer);
 
                     jsonObject.put("connectorResponse", connectorResponse);
                     response = ResponseEntity.ok(jsonObject.toJSONString());
