+++ src/main/java/de/fraunhofer/isst/configmanager/api/controller/ResourceRepresentationController.java
@@ -1,18 +1,12 @@
 package de.fraunhofer.isst.configmanager.api.controller;
 
 
-import de.fraunhofer.iais.eis.Artifact;
-import de.fraunhofer.iais.eis.ArtifactBuilder;
-import de.fraunhofer.iais.eis.IANAMediaTypeBuilder;
-import de.fraunhofer.iais.eis.Language;
-import de.fraunhofer.iais.eis.RepresentationBuilder;
-import de.fraunhofer.iais.eis.RepresentationImpl;
-import de.fraunhofer.iais.eis.ResourceImpl;
+import de.fraunhofer.iais.eis.*;
 import de.fraunhofer.iais.eis.ids.jsonld.Serializer;
 import de.fraunhofer.iais.eis.util.Util;
 import de.fraunhofer.isst.configmanager.api.ResourceRepresentationApi;
 import de.fraunhofer.isst.configmanager.api.service.ConfigModelService;
-import de.fraunhofer.isst.configmanager.api.service.ResourceService;
+import de.fraunhofer.isst.configmanager.api.service.resources.ResourceRepresentationService;
 import de.fraunhofer.isst.configmanager.connector.clients.DefaultResourceClient;
 import de.fraunhofer.isst.configmanager.util.ValidateApiInput;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -42,18 +36,18 @@ import java.net.URI;
 public class ResourceRepresentationController implements ResourceRepresentationApi {
 
     transient ConfigModelService configModelService;
-    transient ResourceService resourceService;
+    transient ResourceRepresentationService resourceRepresentationService;
     transient DefaultResourceClient client;
     transient Serializer serializer;
 
     @Autowired
     public ResourceRepresentationController(final ConfigModelService configModelService,
-                                            final ResourceService resourceService,
+                                            final ResourceRepresentationService resourceRepresentationService,
                                             final DefaultResourceClient client,
                                             final Serializer serializer) {
         this.client = client;
         this.configModelService = configModelService;
-        this.resourceService = resourceService;
+        this.resourceRepresentationService = resourceRepresentationService;
         this.serializer = serializer;
     }
 
@@ -83,7 +77,7 @@ public class ResourceRepresentationController implements ResourceRepresentationA
         if (ValidateApiInput.notValid(resourceId.toString(), sourceType)) {
             response = ResponseEntity.badRequest().body("All validated parameter have undefined as value!");
         } else {
-            if (resourceService.getResources() == null || resourceService.getResources().isEmpty()) {
+            if (resourceRepresentationService.getResources() == null || resourceRepresentationService.getResources().isEmpty()) {
                 response = ResponseEntity.status(HttpStatus.NOT_FOUND).body("Could not find any resources!");
             } else {
                 final var representation = new RepresentationBuilder()
@@ -141,7 +135,7 @@ public class ResourceRepresentationController implements ResourceRepresentationA
                 + filenameExtension + " bytesize: " + bytesize + " sourceType: " + sourceType);
         ResponseEntity<String> response = null;
 
-        final var oldResourceCatalog = (ResourceImpl) resourceService.getResource(resourceId);
+        final var oldResourceCatalog = (ResourceImpl) resourceRepresentationService.getResource(resourceId);
 
         if (oldResourceCatalog != null) {
             final var oldRepresentationId = oldResourceCatalog.getRepresentation().get(0).getId();
@@ -150,7 +144,7 @@ public class ResourceRepresentationController implements ResourceRepresentationA
             if (configModelService.getConfigModel().getAppRoute() == null) {
                 log.info("---- [ResourceRepresentationController updateResourceRepresentation] No AppRoute in ConfigModel!");
             } else {
-                final var oldResourceRoute = (ResourceImpl) resourceService.getResourceInAppRoute(resourceId);
+                final var oldResourceRoute = (ResourceImpl) resourceRepresentationService.getResourceInAppRoute(resourceId);
                 if (oldResourceRoute != null) {
                     oldResourceRoute.setRepresentation(null);
                 }
@@ -187,7 +181,7 @@ public class ResourceRepresentationController implements ResourceRepresentationA
                 }
             }
             // Update the backend connection to the new endpoint
-            resourceService.updateBackendConnection(resourceId, endpointId);
+            resourceRepresentationService.updateBackendConnection(resourceId, endpointId);
 
             try {
                 // Update the resource representation in the dataspace connector
@@ -231,7 +225,7 @@ public class ResourceRepresentationController implements ResourceRepresentationA
         log.info(">> GET /resource/representation representationId: " + representationId);
         ResponseEntity<String> response;
 
-        final var representation = resourceService.getResourceRepresentationInCatalog(representationId);
+        final var representation = resourceRepresentationService.getResourceRepresentationInCatalog(representationId);
 
         if (representation != null) {
             try {
@@ -262,7 +256,7 @@ public class ResourceRepresentationController implements ResourceRepresentationA
         ResponseEntity<String> response = null;
 
         final var representationJson = new JSONObject();
-        for (final var resource : resourceService.getResources()) {
+        for (final var resource : resourceRepresentationService.getResources()) {
             if (resourceId.equals(resource.getId())) {
                 for (final var representation : resource.getRepresentation()) {
                     if (representationId.equals(representation.getId())) {
@@ -300,7 +294,7 @@ public class ResourceRepresentationController implements ResourceRepresentationA
 
         try {
             final var clientResponse = client.deleteResourceRepresentation(resourceId.toString(), representationId.toString());
-            resourceService.deleteResourceRepresentationFromAppRoute(resourceId, representationId);
+            resourceRepresentationService.deleteResourceRepresentationFromAppRoute(resourceId, representationId);
             final var jsonObject = new JSONObject();
             jsonObject.put("connectorResponse", clientResponse);
             jsonObject.put("resourceID", resourceId.toString());
