+++ src/main/java/de/fraunhofer/isst/configmanager/api/service/EndpointService.java
@@ -1,10 +1,6 @@
 package de.fraunhofer.isst.configmanager.api.service;
 
-import de.fraunhofer.iais.eis.BasicAuthenticationBuilder;
-import de.fraunhofer.iais.eis.Endpoint;
-import de.fraunhofer.iais.eis.GenericEndpoint;
-import de.fraunhofer.iais.eis.GenericEndpointBuilder;
-import de.fraunhofer.iais.eis.GenericEndpointImpl;
+import de.fraunhofer.iais.eis.*;
 import de.fraunhofer.isst.configmanager.model.configlists.CustomGenericEndpointList;
 import de.fraunhofer.isst.configmanager.model.configlists.CustomGenericEndpointRepository;
 import de.fraunhofer.isst.configmanager.model.customgenericendpoint.CustomGenericEndpointObject;
