+++ src/main/java/de/fraunhofer/isst/configmanager/model/customgenericendpoint/BackendConfig.java
@@ -7,12 +7,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
-import javax.persistence.Column;
-import javax.persistence.Convert;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
+import javax.persistence.*;
 
 /**
  * The class helps to persist objects inherited from the endpoint from the information model.
