+++ src/main/java/de/fraunhofer/isst/configmanager/util/Utility.java
@@ -1,10 +1,6 @@
 package de.fraunhofer.isst.configmanager.util;
 
-import de.fraunhofer.iais.eis.AppEndpoint;
-import de.fraunhofer.iais.eis.AppEndpointBuilder;
-import de.fraunhofer.iais.eis.AppEndpointType;
-import de.fraunhofer.iais.eis.CustomMediaTypeBuilder;
-import de.fraunhofer.iais.eis.Language;
+import de.fraunhofer.iais.eis.*;
 import de.fraunhofer.iais.eis.util.TypedLiteral;
 import de.fraunhofer.iais.eis.util.Util;
 import lombok.experimental.UtilityClass;
